Added new sub-components to ratings and reviews
