+++ client/src/components/RatingsAndReviews/Ratings.jsx
@@ -1,0 +1,3 @@
+import React from 'react';
+import axios from 'axios';
+import styled from 'styled-components';
